requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiohttp-cors==0.7.0
aioredis==1.3.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
bleach==4.1.0
bsuite==0.3.5
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
chex==0.1.0
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
descartes==1.1.0
distrax==0.1.2
dm-env==1.5
dm-haiku==0.0.6
dm-tree==0.1.6
docopt==0.6.2
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
flatbuffers==2.0
fonttools==4.29.1
frozendict==2.3.0
frozenlist==1.3.0
gast==0.5.3
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.7.1
google-auth==2.6.0
googleapis-common-protos==1.55.0
grpcio==1.44.0
gym==0.22.0
gym-notices==0.0.4
hiredis==2.0.0
idna==3.3
imageio==2.16.0
importlib-metadata==4.11.1
ipykernel==6.12.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jax==0.3.5
jaxlib==0.3.5+cuda11.cudnn82
jedi==0.18.1
Jinja2==3.0.3
jmp==0.0.2
jsonpickle==1.5.2
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mizani==0.7.3
msgpack==1.0.3
multidict==6.0.2
munch==2.5.0
mypy==0.931
mypy-extensions==0.4.3
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.5
networkx==2.6.3
notebook==6.4.8
numpy==1.22.2
opencensus==0.8.0
opencensus-context==0.1.2
opt-einsum==3.3.0
optax==0.1.1
packaging==21.3
palettable==3.3.0
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
plotnine==0.8.0
prometheus-client==0.13.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
pymongo==4.0.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
ray==1.10.0
redis==4.1.4
requests==2.27.1
rlax==0.1.2
rsa==4.8
sacred==0.8.2
scikit-image==0.19.2
scipy==1.8.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
stack-data==0.2.0
statsmodels==0.13.2
tabulate==0.8.9
tensorflow-probability==0.16.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
tifffile==2022.2.9
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wrapt==1.13.3
yarl==1.7.2
zipp==3.7.0
